finalized csv reader

+++ mvv_demo/core/mvv_iostream.h
@@ -12,7 +12,7 @@ float** read_csv(std::string folder_name, std::string file_name);
 
 void save_grayscale_t(std::string folder_name, std::string file_name, short** frame_triangle_grid, int width, int height);
 
-short** read_grayscale_t(std::string folder_name, std::string file_name);
+int** read_grayscale_t(std::string folder_name, std::string file_name);
 
 int io_test();
